✨ Refactered ListView and supported autofocus at the group name fields
